// resource/generated/models/ProfileSettingsRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BarbarianFoodAPI
 * example
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProfileSettingsRequest
 */
export interface ProfileSettingsRequest {
    /**
     * 
     * @type {string}
     * @memberof ProfileSettingsRequest
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileSettingsRequest
     */
    password?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileSettingsRequest
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileSettingsRequest
     */
    surname?: string;
    /**
     * 
     * @type {string}
     * @memberof ProfileSettingsRequest
     */
    token?: string;
}

export function ProfileSettingsRequestFromJSON(json: any): ProfileSettingsRequest {
    return ProfileSettingsRequestFromJSONTyped(json, false);
}

export function ProfileSettingsRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProfileSettingsRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'email': !exists(json, 'email') ? undefined : json['email'],
        'password': !exists(json, 'password') ? undefined : json['password'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'surname': !exists(json, 'surname') ? undefined : json['surname'],
        'token': !exists(json, 'token') ? undefined : json['token'],
    };
}

export function ProfileSettingsRequestToJSON(value?: ProfileSettingsRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'email': value.email,
        'password': value.password,
        'name': value.name,
        'surname': value.surname,
        'token': value.token,
    };
}